fix: resolving error when parsing json content returned by gemini ai

+++ backend/controllers/books.controller.js
@@ -112,7 +112,7 @@ const createBooks = asyncHandler(async (req, res) => {
     summary: aiSummary,
   });
 
-  if (bookSummary.summary.length === 0) {
+  if (aiSummary.length === 0 && bookSummary.summary.length === 0) {
     await Summary.findByIdAndDelete(bookSummary._id);
     await Books.findByIdAndDelete(book._id);
     res.status(406);
